now able to create shared glfw contexts

+++ src/perspectivecamera.cpp
@@ -9,8 +9,8 @@ namespace yavin {
 //==============================================================================
 
 perspectivecamera::perspectivecamera(float fovy, float aspect, float near,
-                                     float far, float vp_x, float vp_y,
-                                     float vp_width, float vp_height)
+                                     float far, size_t vp_x, size_t vp_y,
+                                     size_t vp_width, size_t vp_height)
     : camera{vp_x, vp_y, vp_width, vp_height} {
   set_projection(fovy, aspect, near, far);
 }
@@ -18,7 +18,7 @@ perspectivecamera::perspectivecamera(float fovy, float aspect, float near,
 //------------------------------------------------------------------------------
 
 perspectivecamera::perspectivecamera(float fovy, float aspect, float near,
-                                     float far, float vp_width, float vp_height)
+                                     float far, size_t vp_width, size_t vp_height)
     : camera{vp_width, vp_height} {
   set_projection(fovy, aspect, near, far);
 }
@@ -26,7 +26,7 @@ perspectivecamera::perspectivecamera(float fovy, float aspect, float near,
 //------------------------------------------------------------------------------
 
 perspectivecamera::perspectivecamera(float fovy, float aspect, float near,
-                                     float far, glm::vec4& vp)
+                                     float far, glm::uvec4& vp)
     : camera{vp} {
   set_projection(fovy, aspect, near, far);
 }
@@ -34,7 +34,7 @@ perspectivecamera::perspectivecamera(float fovy, float aspect, float near,
 //------------------------------------------------------------------------------
 
 perspectivecamera::perspectivecamera(float fovy, float aspect, float near,
-                                     float far, glm::vec4&& vp)
+                                     float far, glm::uvec4&& vp)
     : camera{std::move(vp)} {
   set_projection(fovy, aspect, near, far);
 }
@@ -49,8 +49,8 @@ void perspectivecamera::set_projection(float fovy, float aspect, float near,
 //------------------------------------------------------------------------------
 
 void perspectivecamera::set_projection(float fovy, float aspect, float near,
-                                       float far, float vp_width,
-                                       float vp_height) {
+                                       float far, size_t vp_width,
+                                       size_t vp_height) {
   set_projection(fovy, aspect, near, far);
   set_viewport(vp_width, vp_height);
 }
@@ -58,8 +58,8 @@ void perspectivecamera::set_projection(float fovy, float aspect, float near,
 //------------------------------------------------------------------------------
 
 void perspectivecamera::set_projection(float fovy, float aspect, float near,
-                                       float far, float vp_x, float vp_y,
-                                       float vp_width, float vp_height) {
+                                       float far, size_t vp_x, size_t vp_y,
+                                       size_t vp_width, size_t vp_height) {
   set_projection(fovy, aspect, near, far);
   set_viewport(vp_x, vp_y, vp_width, vp_height);
 }
